+++ Microsoft.Toolkit.Uwp.UI.Controls.DataGrid/Utilities/Extensions.cs
@@ -190,7 +190,7 @@ namespace Microsoft.Toolkit.Uwp.UI.Utilities
         {
             if (element.DesiredSize.Height == 0)
             {
-                element.Measure(new Size(double.PositiveInfinity, double.PositiveInfinity));
+                element.Measure(new Size(float.PositiveInfinity, float.PositiveInfinity));
             }
 
             return element;
